requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.15.0
mysql-connector-python>=8.1.0
sqlalchemy>=2.0.0
PyYAML>=6.0
bcrypt>=4.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Microsoft Entra ID Authentication
streamlit-oauth>=0.1.0
requests>=2.31.0
PyJWT>=2.8.0
cryptography>=41.0.0